avoid the usage of runtime to access the cleaner
